+++ FMS.Application/ModelsDTOs/FMS/FuelRefil/FuelRefilDTO.cs
@@ -12,9 +12,9 @@ namespace FMS.Application.ModelsDTOs.FMS.FuelRefil
 
     public DateTime? Date { get; set; }
 
-    public int? PreviousMeterReading { get; set; }
+    public decimal? PreviousMeterReading { get; set; }
 
-    public int? CurrentMeterReading { get; set; }
+    public decimal? CurrentMeterReading { get; set; }
 
     public int SiteId { get; set; }
